requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.2.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
nest-asyncio==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==25.1.2
requests==2.31.0
rich==13.7.1
setuptools==69.2.0
six==1.16.0
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
